Turn front page into a redirect and add years 2024 (live) and 2025 (hidden).
