+++ gatsby-config.js
@@ -4,6 +4,490 @@ const path = require(`path`);
 
 module.exports = {
   siteMetadata: {
+    cvpr2025: {
+      organizers: [
+        {
+          challenge: false,
+          imageId: "adeFamoti",
+          name: "Ade Famoti",
+          oc: true,
+          lo: false,
+          organization: "Microsoft",
+          sab: false,
+          site: "https://www.microsoft.com/en-us/research/people/adfamoti/",
+        },
+        {
+          challenge: false,
+          imageId: "alexander",
+          name: "Alexander Toshev",
+          oc: false,
+          lo: false,
+          organization: "Apple",
+          sab: true,
+          site: "https://sites.google.com/view/alextoshev",
+        },
+        {
+          challenge: false,
+          imageId: "andreyKolobov",
+          name: "Andrey Kolobov",
+          oc: false,
+          lo: false,
+          organization: "Microsoft",
+          sab: true,
+          site: "https://www.microsoft.com/en-us/research/people/akolobov/",
+        },
+        {
+          challenge: true,
+          imageId: "angel",
+          name: "Angel X. Chang",
+          oc: true,
+          lo: false,
+          organization: "SFU",
+          sab: false,
+          site: "https://angelxuanchang.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "ani",
+          name: "Aniruddha Kembhavi",
+          oc: false,
+          lo: false,
+          organization: "AI2, UW",
+          sab: true,
+          site: "https://anikem.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "anthony",
+          name: "Anthony Francis",
+          oc: true,
+          lo: true,
+          organization: "Logical Robotics",
+          sab: false,
+          site: "https://logicalrobotics.com/",
+        },
+        {
+          challenge: true,
+          imageId: "baoxiongJia",
+          name: "Baoxiong Jia",
+          oc: false,
+          lo: false,
+          organization: "BIGAI",
+          sab: false,
+          site: "https://scholar.google.com/citations?user=qIBUK6sAAAAJ&hl=en",
+        },        
+        {
+          challenge: true,
+          imageId: "changan",
+          name: "Changan Chen",
+          oc: true,
+          lo: false,
+          organization: "UT Austin",
+          sab: false,
+          site: "https://changan.io/",
+        },
+        {
+          challenge: false,
+          imageId: "chengshu",
+          name: "Chengshu Li",
+          oc: true,
+          lo: false,
+          organization: "Stanford",
+          sab: false,
+          site: "https://www.chengshuli.me/",
+        },
+        {
+          challenge: false,
+          imageId: "chris",
+          name: "Chris Paxton",
+          oc: false,
+          lo: false,
+          organization: "Meta AI",
+          sab: false,
+          site: "https://cpaxton.github.io/about/",
+        },
+        {
+          challenge: true,
+          imageId: "chuang",
+          name: "Chuang Gan",
+          oc: false,
+          lo: false,
+          organization: "IBM, MIT",
+          sab: false,
+          site: "https://people.csail.mit.edu/ganchuang/",
+        },
+        {
+          challenge: false,
+          imageId: "claudia",
+          name: "Claudia Pérez D\u2019Arpino",
+          oc: true,
+          lo: true,
+          organization: "NVIDIA",
+          sab: false,
+          site: "https://ai.stanford.edu/~cdarpino/",
+        },
+        {
+          challenge: true,
+          imageId: "davidH",
+          name: "David Hall",
+          oc: true,
+          lo: false,
+          organization: "CSIRO",
+          sab: false,
+          site: "https://sites.google.com/view/davidhallcv/home",
+        },
+        {
+          challenge: false,
+          imageId: "devon",
+          name: "Devon Hjelm",
+          oc: true,
+          lo: false,
+          organization: "Apple",
+          sab: false,
+          site: "https://rdevon.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "dhruv",
+          name: "Dhruv Batra",
+          oc: false,
+          lo: false,
+          organization: "GaTech, Meta AI",
+          sab: true,
+          site: "https://faculty.cc.gatech.edu/~dbatra/",
+        },
+        {
+          challenge: true,
+          imageId: "fanbo",
+          name: "Fanbo Xiang",
+          oc: false,
+          lo: false,
+          organization: "UCSD",
+          sab: false,
+          site: "https://www.fbxiang.com/",
+        },
+        {
+          challenge: false,
+          imageId: "german",
+          name: "German Ros",
+          oc: false,
+          lo: false,
+          organization: "NVIDIA",
+          sab: true,
+          site: "https://germanros.net/",
+        },
+        {
+          challenge: true,
+          imageId: "haoDong",
+          name: "Hao Dong",
+          oc: false,
+          lo: false,
+          organization: "PKU",
+          sab: false,
+          site: "https://zsdonghao.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "jiayuan",
+          name: "Jiayuan Gu",
+          oc: false,
+          lo: false,
+          organization: "UCSD",
+          sab: false,
+          site: "https://cseweb.ucsd.edu/~jigu/",
+        },
+        {
+          challenge: true,
+          imageId: "jiangyongHuang",
+          name: "Jiangyong Huang",
+          oc: false,
+          lo: false,
+          organization: "Peking U",
+          sab: false,
+          site: "https://huangjy-pku.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "joanne",
+          name: "Joanne Truong",
+          oc: false,
+          lo: false,
+          organization: "GaTech",
+          sab: true,
+          site: "https://www.joannetruong.com/",
+        },
+        {
+          challenge: false,
+          imageId: "joelJang",
+          name: "Joel Jang",
+          oc: true,
+          lo: false,
+          organization: "U Washington",
+          sab: false,
+          site: "https://joeljang.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "lamberto",
+          name: "Lamberto Ballan",
+          oc: true,
+          lo: false,
+          organization: "U Padova",
+          sab: false,
+          site: "http://www.lambertoballan.net/",
+        },
+        {
+          challenge: true,
+          imageId: "luca",
+          name: "Luca Weihs",
+          oc: true,
+          lo: true,
+          organization: "AI2",
+          sab: false,
+          site: "https://lucaweihs.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "manolis",
+          name: "Manolis Savva",
+          oc: false,
+          lo: false,
+          organization: "SFU",
+          sab: true,
+          site: "https://msavva.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "matt",
+          name: "Matt Deitke",
+          oc: true,
+          lo: false,
+          organization: "AI2, UW",
+          sab: false,
+          site: "https://mattdeitke.com/",
+        },
+        {
+          challenge: false,
+          imageId: "mike",
+          name: "Mike Roberts",
+          oc: true,
+          lo: false,
+          organization: "Intel",
+          sab: false,
+          site: "https://mikeroberts3000.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "naoki",
+          name: "Naoki Yokoyama",
+          oc: true,
+          lo: false,
+          organization: "GaTech",
+          sab: false,
+          site: "http://naoki.io/",
+        },
+        {
+          challenge: true,
+          imageId: "oleksandr",
+          name: "Oleksandr Maksymets",
+          oc: true,
+          lo: false,
+          organization: "Meta AI",
+          sab: false,
+          site: "https://research.facebook.com/people/maksymets-oleksandr/",
+        },
+        {
+          challenge: true,
+          imageId: "ram",
+          name: "Ram Ramrakhya",
+          oc: true,
+          lo: false,
+          organization: "Gatech",
+          sab: false,
+          site: "https://ram81.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "ranGong",
+          name: "Ran Gong",
+          oc: true,
+          lo: false,
+          organization: "UCLA",
+          sab: false,
+          site: "https://nikepupu.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "richardHeBai",
+          name: "Richard He Bai",
+          oc: false,
+          lo: false,
+          organization: "Apple",
+          sab: false,
+          site: "https://richardbaihe.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "rin",
+          name: "Rin Metcalf",
+          oc: true,
+          lo: false,
+          organization: "Apple",
+          sab: false,
+          site: "https://scholar.google.com/citations?user=V7baeTMAAAAJ",
+        },
+        {
+          challenge: false,
+          imageId: "roberto",
+          name: "Roberto Mart\u00edn-Mart\u00edn",
+          oc: false,
+          lo: false,
+          organization: "Stanford",
+          sab: true,
+          site: "https://robertomartinmartin.com/",
+        },
+        {
+          challenge: true,
+          imageId: "roozbeh",
+          name: "Roozbeh Mottaghi",
+          oc: false,
+          lo: false,
+          organization: "FAIR, UW",
+          sab: true,
+          site: "https://roozbehm.info/",
+        },
+        {
+          challenge: true,
+          imageId: "siyuanHuang",
+          name: "Siyuan Huang",
+          oc: false,
+          lo: false,
+          organization: "BIGAI",
+          sab: false,
+          site: "https://siyuanhuang.com/",
+        },
+        {
+          challenge: true,
+          imageId: "soniaR",
+          name: "Sonia Raychaudhuri",
+          oc: false,
+          lo: false,
+          organization: "SFU",
+          sab: false,
+          site: "https://sonia-raychaudhuri.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "sorean",
+          name: "S\u00f6ren Pirk",
+          oc: true,
+          lo: false,
+          organization: "Kiel University",
+          sab: false,
+          site: "https://www.pirk.io",
+        },
+        {
+          challenge: true,
+          imageId: "stone",
+          name: "Stone Tao",
+          oc: false,
+          lo: false,
+          organization: "UCSD",
+          sab: false,
+          site: "https://www.stoneztao.com/",
+        },
+        {
+          challenge: true,
+          imageId: "tommaso",
+          name: "Tommaso Campari",
+          oc: false,
+          lo: false,
+          organization: "SFU, UNIPD",
+          sab: false,
+          site: "https://www.tommasocampari.com",
+        },
+        {
+          challenge: true,
+          imageId: "unnat",
+          name: "Unnat Jain",
+          oc: false,
+          lo: false,
+          organization: "UIUC",
+          sab: false,
+          site: "https://unnat.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "xiaofeng",
+          name: "Xiaofeng Gao",
+          oc: false,
+          lo: false,
+          organization: "Amazon",
+          sab: false,
+          site: "https://xfgao.github.io/",
+        },
+        {
+          challenge: true,
+          imageId: "yangLiu",
+          name: "Yang Liu",
+          oc: false,
+          lo: false,
+          organization: "SRC-B",
+          sab: false,
+          site: "",
+        },
+        {
+          challenge: true,
+          imageId: "yonatan",
+          name: "Yonatan Bisk",
+          oc: true,
+          lo: false,
+          organization: "CMU",
+          sab: false,
+          site: "https://yonatanbisk.com/",
+        },
+        {
+          challenge: true,
+          imageId: "zhuoqunXu",
+          name: "Zhuoqun Xu",
+          oc: false,
+          lo: false,
+          organization: "PRS",
+          sab: false,
+          site: "https://zhqunxu.github.io/",
+        },
+        {
+          challenge: false,
+          imageId: "stevieBathiche",
+          name: "Stevie Bathiche",
+          oc: false,
+          lo: false,
+          organization: "Microsoft",
+          sab: false,
+          site: "https://www.microsoft.com/applied-sciences/people/steven-bathiche",
+        },
+        {
+          challenge: false,
+          imageId: "ashleyLlorens",
+          name: "Ashley Llorens",
+          oc: false,
+          lo: false,
+          organization: "Microsoft",
+          sab: false,
+          site: "https://www.microsoft.com/en-us/research/people/allorens/",
+        },
+        {
+          challenge: false,
+          imageId: "geordieRose",
+          name: "Geordie Rose",
+          oc: false,
+          lo: false,
+          organization: "Sanctuary AI",
+          sab: false,
+          site: "https://sanctuary.ai/bio-geordie-rose/",
+        },
+      ]
+    },
     cvpr2024: {
       organizers: [
         {
